refactor(dispatch): probe seam lives in its own module
A same-module helper on safe-spawn.ts would call the module's own
unmocked safeSpawnAsync binding, so every availability test that fakes
the process boundary with vi.mock("clients/safe-spawn.js") would have
spawned for real (23 reds in availability-latching-siblings). One module
hop keeps every existing double production-faithful.

Refs #2894
